refactor(registration): split large form into modular components (personal info, profession, zone, preferences & submit)
