+++ src/components/registration/success/RegistrationSuccess.tsx
